+++ src/main/java/com/commercehub/gradle/plugin/avro/AvroExtension.java
@@ -15,12 +15,13 @@
  */
 package com.commercehub.gradle.plugin.avro;
 
-import java.util.List;
-import java.util.Map;
 import org.apache.avro.Conversion;
 import org.apache.avro.LogicalTypes;
+import org.gradle.api.provider.ListProperty;
+import org.gradle.api.provider.MapProperty;
 import org.gradle.api.provider.Property;
 
+@SuppressWarnings("unused")
 public interface AvroExtension {
     Property<String> getOutputCharacterEncoding();
     Property<String> getStringType();
@@ -31,12 +32,8 @@ public interface AvroExtension {
     Property<Boolean> isGettersReturnOptional();
     Property<Boolean> isEnableDecimalLogicalType();
     Property<String> getDateTimeLogicalType();
-    // When we require Gradle 5.1+, we could use MapProperty here.
-    @SuppressWarnings("rawtypes")
-    Property<Map> getLogicalTypeFactories();
-    // When we require Gradle 4.5+, we could use ListProperty here.
-    @SuppressWarnings("rawtypes")
-    Property<List> getCustomConversions();
+    MapProperty<String, Class<? extends LogicalTypes.LogicalTypeFactory>> getLogicalTypeFactories();
+    ListProperty<Class<? extends Conversion<?>>> getCustomConversions();
     AvroExtension logicalTypeFactory(String typeName, Class<? extends LogicalTypes.LogicalTypeFactory> typeFactoryClass);
     AvroExtension customConversion(Class<? extends Conversion<?>> conversionClass);
 }