+++ src/main/java/com/commercehub/gradle/plugin/avro/Constants.java
@@ -18,9 +18,12 @@ package com.commercehub.gradle.plugin.avro;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import org.apache.avro.Conversion;
+import org.apache.avro.LogicalTypes;
 import org.apache.avro.compiler.specific.SpecificCompiler;
 import org.apache.avro.compiler.specific.SpecificCompiler.FieldVisibility;
 import org.apache.avro.generic.GenericData.StringType;
+import org.gradle.api.reflect.TypeOf;
 
 /**
  * Various constants needed by the plugin.
@@ -38,10 +41,8 @@ class Constants {
     static final boolean DEFAULT_GETTERS_RETURN_OPTIONAL = false;
     static final boolean DEFAULT_ENABLE_DECIMAL_LOGICAL_TYPE = true;
     static final String DEFAULT_DATE_TIME_LOGICAL_TYPE = SpecificCompiler.DateTimeLogicalTypeImplementation.DEFAULT.name();
-    @SuppressWarnings("rawtypes")
-    static final Map<String, Class> DEFAULT_LOGICAL_TYPE_FACTORIES = Collections.emptyMap();
-    @SuppressWarnings("rawtypes")
-    static final List<Class> DEFAULT_CUSTOM_CONVERSIONS = Collections.emptyList();
+    static final Map<String, Class<? extends LogicalTypes.LogicalTypeFactory>> DEFAULT_LOGICAL_TYPE_FACTORIES = Collections.emptyMap();
+    static final List<Class<? extends Conversion<?>>> DEFAULT_CUSTOM_CONVERSIONS = Collections.emptyList();
 
     static final String SCHEMA_EXTENSION = "avsc";
     static final String PROTOCOL_EXTENSION = "avpr";
@@ -55,4 +56,10 @@ class Constants {
     static final String OPTION_FIELD_VISIBILITY = "fieldVisibility";
     static final String OPTION_STRING_TYPE = "stringType";
     static final String OPTION_DATE_TIME_LOGICAL_TYPE = "dateTimeLogicalType";
+
+    static final TypeOf<Class<? extends LogicalTypes.LogicalTypeFactory>> LOGICAL_TYPE_FACTORY_TYPE =
+        new TypeOf<Class<? extends LogicalTypes.LogicalTypeFactory>>() { };
+
+    static final TypeOf<Class<? extends Conversion<?>>> CONVERSION_TYPE =
+        new TypeOf<Class<? extends Conversion<?>>>() { };
 }