+++ src/main/java/com/commercehub/gradle/plugin/avro/GradleFeatures.java
@@ -25,24 +25,6 @@ enum GradleFeatures {
             return GradleVersion.current().compareTo(GradleVersions.v5_2) >= 0;
         }
     },
-    propertyConventions() {
-        @Override
-        boolean isSupported() {
-            return GradleVersion.current().compareTo(GradleVersions.v5_1) >= 0;
-        }
-    },
-    objectFactoryDirectoryProperty() {
-        @Override
-        boolean isSupported() {
-            return GradleVersion.current().compareTo(GradleVersions.v5_0) >= 0;
-        }
-    },
-    projectLayoutConfigurableFiles() {
-        @Override
-        boolean isSupported() {
-            return GradleVersion.current().compareTo(GradleVersions.v4_8) >= 0;
-        }
-    },
     objectFactoryFileCollection() {
         @Override
         boolean isSupported() {